Split entities into 3 files, constants, descriptors and entities and code reformating
