+++ tests/test_entities.py
@@ -1,268 +1,15 @@
+from sys import version_info
 from unittest import TestCase
 from xml.etree import ElementTree
-from sys import version_info
-from io import BytesIO
 
-from genologics.lims import Lims
 from genologics.entities import StepActions, Researcher, Artifact, \
-    Step, StepPlacements, Container, Stage, ReagentKit, ReagentLot,\
-    StringDescriptor, StringAttributeDescriptor, StringListDescriptor, \
-    StringDictionaryDescriptor, IntegerDescriptor, BooleanDescriptor, UdfDictionary, EntityDescriptor
+    Step, StepPlacements, Container, Stage, ReagentKit, ReagentLot
+from genologics.lims import Lims
 
 if version_info.major == 2:
     from mock import patch, Mock
 else:
     from unittest.mock import patch, Mock
-
-
-class TestDescriptor(TestCase):
-
-    def _make_desc(self, klass, *args, **kwargs):
-        return klass(*args, **kwargs)
-
-    def _tostring(self, e):
-        outfile = BytesIO()
-        ElementTree.ElementTree(e).write(outfile, encoding='utf-8', xml_declaration=True)
-        return outfile.getvalue()
-
-
-class TestStringDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="utf-8"?>
-<test-entry>
-<name>test name</name>
-</test-entry>
-""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringDescriptor, 'name')
-        assert sd.__get__(self.instance, None) == "test name"
-
-    def test__set__(self):
-        sd = self._make_desc(StringDescriptor, 'name')
-        sd.__set__(self.instance, "new test name")
-        assert self.et.find('name').text == "new test name"
-
-    def test_create(self):
-        instance_new = Mock(root= ElementTree.Element('test-entry'))
-        sd = self._make_desc(StringDescriptor, 'name')
-        sd.__set__(instance_new, "test name")
-        assert instance_new.root.find('name').text == 'test name'
-
-class TestIntegerDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<count>32</count>
-</test-entry>
-""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(IntegerDescriptor, 'count')
-        assert sd.__get__(self.instance, None) == 32
-
-    def test__set__(self):
-        sd = self._make_desc(IntegerDescriptor, 'count')
-        sd.__set__(self.instance, 23)
-        assert self.et.find('count').text == '23'
-        sd.__set__(self.instance, '23')
-        assert self.et.find('count').text == '23'
-
-
-    def test_create(self):
-        instance_new = Mock(root= ElementTree.Element('test-entry'))
-        sd = self._make_desc(IntegerDescriptor, 'count')
-        sd.__set__(instance_new, 23)
-        assert instance_new.root.find('count').text == '23'
-
-class TestBooleanDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<istest>true</istest>
-</test-entry>
-""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        bd = self._make_desc(BooleanDescriptor, 'istest')
-        assert bd.__get__(self.instance, None) == True
-
-    def test__set__(self):
-        bd = self._make_desc(BooleanDescriptor, 'istest')
-        bd.__set__(self.instance, False)
-        assert self.et.find('istest').text == 'false'
-        bd.__set__(self.instance, 'true')
-        assert self.et.find('istest').text == 'true'
-
-    def test_create(self):
-        instance_new = Mock(root= ElementTree.Element('test-entry'))
-        bd = self._make_desc(BooleanDescriptor, 'istest')
-        bd.__set__(instance_new, True)
-        assert instance_new.root.find('istest').text == 'true'
-
-
-class TestEntityDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<artifact uri="http://testgenologics.com:4040/api/v2/artifacts/a1"></artifact>
-</test-entry>
-""")
-        self.lims = Lims('http://testgenologics.com:4040', username='test', password='password')
-        self.a1 = Artifact(self.lims,id='a1')
-        self.a2 = Artifact(self.lims,id='a2')
-        self.instance = Mock(root=self.et, lims=self.lims)
-
-    def test__get__(self):
-        ed = self._make_desc(EntityDescriptor, 'artifact', Artifact)
-        assert ed.__get__(self.instance, None) == self.a1
-
-    def test__set__(self):
-        ed = self._make_desc(EntityDescriptor, 'artifact', Artifact)
-        ed.__set__(self.instance, self.a2)
-        assert self.et.find('artifact').attrib['uri'] == 'http://testgenologics.com:4040/api/v2/artifacts/a2'
-
-    def test_create(self):
-        instance_new = Mock(root= ElementTree.Element('test-entry'))
-        ed = self._make_desc(EntityDescriptor, 'artifact', Artifact)
-        ed.__set__(instance_new, self.a1)
-        assert instance_new.root.find('artifact').attrib['uri'] == 'http://testgenologics.com:4040/api/v2/artifacts/a1'
-
-
-class TestStringAttributeDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry name="test name">
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringAttributeDescriptor, 'name')
-        assert sd.__get__(self.instance, None) == "test name"
-
-
-class TestStringListDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<test-subentry>A01</test-subentry>
-<test-subentry>B01</test-subentry>
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringListDescriptor, 'test-subentry')
-        assert sd.__get__(self.instance, None) == ['A01', 'B01']
-
-
-class TestStringDictionaryDescriptor(TestDescriptor):
-
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry>
-<test-subentry>
-<test-firstkey/>
-<test-secondkey>second value</test-secondkey>
-</test-subentry>
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-
-    def test__get__(self):
-        sd = self._make_desc(StringDictionaryDescriptor, 'test-subentry')
-        res = sd.__get__(self.instance, None)
-        assert type(res) == dict
-        self.assertIsNone(res['test-firstkey'])
-        assert res['test-secondkey'] == 'second value'
-
-
-class TestUdfDictionary(TestCase):
-    def setUp(self):
-        self.et = ElementTree.fromstring("""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<test-entry xmlns:udf="http://genologics.com/ri/userdefined">
-<udf:field type="String" name="test">stuff</udf:field>
-<udf:field type="Numeric" name="how much">42</udf:field>
-</test-entry>""")
-        self.instance = Mock(root=self.et)
-        self.dict1 = UdfDictionary(self.instance)
-        self.dict2 = UdfDictionary(self.instance, udt=True)
-
-    def _get_udf_value(self, udf_dict, key):
-        for e in udf_dict._elems:
-            if e.attrib['name'] != key:
-                continue
-            else:
-                return e.text
-
-    def test_get_udt(self):
-        pass
-
-    def test_set_udt(self):
-        pass
-
-    def test__update_elems(self):
-        pass
-
-    def test__prepare_lookup(self):
-        pass
-
-    def test___contains__(self):
-        pass
-
-    def test___getitem__(self):
-        pass
-
-    def test___setitem__(self):
-        assert self._get_udf_value(self.dict1, 'test') == 'stuff'
-        self.dict1.__setitem__('test', 'other')
-        assert self._get_udf_value(self.dict1, 'test') == 'other'
-
-        assert self._get_udf_value(self.dict1, 'how much') == '42'
-        self.dict1.__setitem__('how much', 21)
-        assert self._get_udf_value(self.dict1, 'how much') == '21'
-
-        self.assertRaises(TypeError, self.dict1.__setitem__, 'how much', '433')
-
-        #I'm not sure if this is the expected behaviour
-        self.dict1.__setitem__('how much', None)
-        assert self._get_udf_value(self.dict1, 'how much') == b'None'
-
-    def test___setitem__unicode(self):
-        assert self._get_udf_value(self.dict1, 'test') == 'stuff'
-        self.dict1.__setitem__('test', u'unicode')
-        assert self._get_udf_value(self.dict1, 'test') == 'unicode'
-
-        self.dict1.__setitem__(u'test', 'unicode2')
-        assert self._get_udf_value(self.dict1, 'test') == 'unicode2'
-
-
-
-    def test___delitem__(self):
-        pass
-
-    def test_items(self):
-        pass
-
-    def test_clear(self):
-        pass
-
-    def test___iter__(self):
-        pass
-
-    def test___next__(self):
-        pass
-
-    def test_get(self):
-        pass
-
 
 url = 'http://testgenologics.com:4040'
 
@@ -369,7 +116,6 @@ generic_step_actions_xml = """<stp:actions xmlns:stp="http://genologics.com/ri/s
   </escalation>
 </stp:actions>"""
 
-
 generic_step_actions_no_escalation_xml = """<stp:actions xmlns:stp="http://genologics.com/ri/step" uri="...">
   <step rel="..." uri="{url}/steps/s1">
   </step>
@@ -380,8 +126,8 @@ generic_step_actions_no_escalation_xml = """<stp:actions xmlns:stp="http://genol
   </next-actions>
 </stp:actions>"""
 
-class TestEntities(TestCase):
 
+class TestEntities(TestCase):
     def test_pass(self):
         pass
 
@@ -394,8 +140,9 @@ def elements_equal(e1, e2):
     if len(e1) != len(e2): return False
     return all(elements_equal(c1, c2) for c1, c2 in zip(e1, e2))
 
+
 class TestEntities(TestCase):
-    dummy_xml="""<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+    dummy_xml = """<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
     <dummy></dummy>"""
 
     def setUp(self):
@@ -411,8 +158,8 @@ class TestStepActions(TestEntities):
 
     def test_escalation(self):
         s = StepActions(uri=self.lims.get_uri('steps', 'step_id', 'actions'), lims=self.lims)
-        with patch('requests.Session.get',return_value=Mock(content = self.step_actions_xml, status_code=200)),\
-             patch('requests.post', return_value=Mock(content = self.dummy_xml, status_code=200)):
+        with patch('requests.Session.get', return_value=Mock(content=self.step_actions_xml, status_code=200)), \
+             patch('requests.post', return_value=Mock(content=self.dummy_xml, status_code=200)):
             r = Researcher(uri='http://testgenologics.com:4040/researchers/r1', lims=self.lims)
             a = Artifact(uri='http://testgenologics.com:4040/artifacts/r1', lims=self.lims)
             expected_escalation = {
@@ -426,7 +173,8 @@ class TestStepActions(TestEntities):
 
     def test_next_actions(self):
         s = StepActions(uri=self.lims.get_uri('steps', 'step_id', 'actions'), lims=self.lims)
-        with patch('requests.Session.get',return_value=Mock(content = self.step_actions_no_escalation_xml, status_code=200)):
+        with patch('requests.Session.get',
+                   return_value=Mock(content=self.step_actions_no_escalation_xml, status_code=200)):
             step1 = Step(self.lims, uri='http://testgenologics.com:4040/steps/s1')
             step2 = Step(self.lims, uri='http://testgenologics.com:4040/steps/s2')
             artifact = Artifact(self.lims, uri='http://testgenologics.com:4040/artifacts/a1')
@@ -436,19 +184,18 @@ class TestStepActions(TestEntities):
 
 
 class TestStepPlacements(TestEntities):
-
     original_step_placements_xml = generic_step_placements_xml.format(url=url, container="c1", loc1='1:1', loc2='2:1')
     modloc_step_placements_xml = generic_step_placements_xml.format(url=url, container="c1", loc1='3:1', loc2='4:1')
     modcont_step_placements_xml = generic_step_placements_xml.format(url=url, container="c2", loc1='1:1', loc2='1:1')
 
-
     def test_get_placements_list(self):
         s = StepPlacements(uri=self.lims.get_uri('steps', 's1', 'placements'), lims=self.lims)
-        with patch('requests.Session.get',return_value=Mock(content = self.original_step_placements_xml, status_code=200)):
+        with patch('requests.Session.get',
+                   return_value=Mock(content=self.original_step_placements_xml, status_code=200)):
             a1 = Artifact(uri='http://testgenologics.com:4040/artifacts/a1', lims=self.lims)
             a2 = Artifact(uri='http://testgenologics.com:4040/artifacts/a2', lims=self.lims)
             c1 = Container(uri='http://testgenologics.com:4040/containers/c1', lims=self.lims)
-            expected_placements = [[a1,(c1,'1:1')], [a2,(c1,'2:1')]]
+            expected_placements = [[a1, (c1, '1:1')], [a2, (c1, '2:1')]]
             assert s.get_placement_list() == expected_placements
 
     def test_set_placements_list(self):
@@ -458,8 +205,9 @@ class TestStepPlacements(TestEntities):
         c2 = Container(uri='http://testgenologics.com:4040/containers/c2', lims=self.lims)
 
         s = StepPlacements(uri=self.lims.get_uri('steps', 's1', 'placements'), lims=self.lims)
-        with patch('requests.Session.get',return_value=Mock(content = self.original_step_placements_xml, status_code=200)):
-            new_placements = [[a1,(c1,'3:1')], [a2,(c1,'4:1')]]
+        with patch('requests.Session.get',
+                   return_value=Mock(content=self.original_step_placements_xml, status_code=200)):
+            new_placements = [[a1, (c1, '3:1')], [a2, (c1, '4:1')]]
             s.set_placement_list(new_placements)
             assert elements_equal(s.root, ElementTree.fromstring(self.modloc_step_placements_xml))
 
@@ -469,33 +217,29 @@ class TestStepPlacements(TestEntities):
         c2 = Container(uri='http://testgenologics.com:4040/containers/c2', lims=self.lims)
 
         s = StepPlacements(uri=self.lims.get_uri('steps', 's1', 'placements'), lims=self.lims)
-        with patch('requests.Session.get',return_value=Mock(content = self.original_step_placements_xml, status_code=200)):
-            new_placements = [[a1,(c2,'1:1')], [a2,(c2,'1:1')]]
+        with patch('requests.Session.get',
+                   return_value=Mock(content=self.original_step_placements_xml, status_code=200)):
+            new_placements = [[a1, (c2, '1:1')], [a2, (c2, '1:1')]]
             s.set_placement_list(new_placements)
             assert elements_equal(s.root, ElementTree.fromstring(self.modcont_step_placements_xml))
 
 
-
-
 class TestArtifacts(TestEntities):
-
     root_artifact_xml = generic_artifact_xml.format(url=url)
 
     def test_input_artifact_list(self):
         a = Artifact(uri=self.lims.get_uri('artifacts', 'a1'), lims=self.lims)
-        with patch('requests.Session.get',return_value=Mock(content = self.root_artifact_xml, status_code=200)):
+        with patch('requests.Session.get', return_value=Mock(content=self.root_artifact_xml, status_code=200)):
             assert a.input_artifact_list() == []
 
     def test_workflow_stages_and_statuses(self):
         a = Artifact(uri=self.lims.get_uri('artifacts', 'a1'), lims=self.lims)
         expected_wf_stage = [
-            (Stage(self.lims, uri = url + '/api/v2/configuration/workflows/1/stages/2'), 'QUEUED', 'Test workflow s2'),
-            (Stage(self.lims, uri = url + '/api/v2/configuration/workflows/1/stages/1'), 'COMPLETE', 'Test workflow s1')
+            (Stage(self.lims, uri=url + '/api/v2/configuration/workflows/1/stages/2'), 'QUEUED', 'Test workflow s2'),
+            (Stage(self.lims, uri=url + '/api/v2/configuration/workflows/1/stages/1'), 'COMPLETE', 'Test workflow s1')
         ]
-        with patch('requests.Session.get',return_value=Mock(content = self.root_artifact_xml, status_code=200)):
+        with patch('requests.Session.get', return_value=Mock(content=self.root_artifact_xml, status_code=200)):
             assert a.workflow_stages_and_statuses == expected_wf_stage
-
-
 
 
 class TestReagentKits(TestEntities):
@@ -504,16 +248,17 @@ class TestReagentKits(TestEntities):
 
     def test_parse_entity(self):
         r = ReagentKit(uri=self.lims.get_uri('reagentkits', 'r1'), lims=self.lims)
-        with patch('requests.Session.get', return_value=Mock(content = self.reagentkit_xml, status_code=200)):
+        with patch('requests.Session.get', return_value=Mock(content=self.reagentkit_xml, status_code=200)):
             assert r.name == 'regaentkitname'
             assert r.supplier == 'reagentProvider'
             assert r.website == 'www.reagentprovider.com'
             assert r.archived == False
 
     def test_create_entity(self):
-        with patch('genologics.lims.requests.post', return_value=Mock(content = self.reagentkit_xml, status_code=201)):
-            r = ReagentKit.create(self.lims, name='regaentkitname', supplier='reagentProvider', website='www.reagentprovider.com', archived=False)
-        self.assertRaises(TypeError,ReagentKit.create,self.lims, error='test')
+        with patch('genologics.lims.requests.post', return_value=Mock(content=self.reagentkit_xml, status_code=201)):
+            r = ReagentKit.create(self.lims, name='regaentkitname', supplier='reagentProvider',
+                                  website='www.reagentprovider.com', archived=False)
+        self.assertRaises(TypeError, ReagentKit.create, self.lims, error='test')
 
 
 class TestReagentLots(TestEntities):
@@ -522,23 +267,24 @@ class TestReagentLots(TestEntities):
 
     def test_parse_entity(self):
         l = ReagentLot(uri=self.lims.get_uri('reagentkits', 'r1'), lims=self.lims)
-        with patch('requests.Session.get', return_value=Mock(content = self.reagentlot_xml, status_code=200)):
+        with patch('requests.Session.get', return_value=Mock(content=self.reagentlot_xml, status_code=200)):
             assert l.uri
             assert l.name == 'kitname'
             assert l.lot_number == '100'
             assert l.status == 'ARCHIVED'
 
     def test_create_entity(self):
-        with patch('requests.Session.get', return_value=Mock(content = self.reagentkit_xml, status_code=200)):
+        with patch('requests.Session.get', return_value=Mock(content=self.reagentkit_xml, status_code=200)):
             r = ReagentKit(uri=self.lims.get_uri('reagentkits', 'r1'), lims=self.lims)
-        with patch('genologics.lims.requests.post', return_value=Mock(content = self.reagentlot_xml, status_code=201)) as patch_post:
+        with patch('genologics.lims.requests.post',
+                   return_value=Mock(content=self.reagentlot_xml, status_code=201)) as patch_post:
             l = ReagentLot.create(
-                self.lims,
-                reagent_kit=r,
-                name='kitname',
-                lot_number='100',
-                expiry_date='2020-05-01',
-                status='ACTIVE'
+                    self.lims,
+                    reagent_kit=r,
+                    name='kitname',
+                    lot_number='100',
+                    expiry_date='2020-05-01',
+                    status='ACTIVE'
             )
             assert l.uri
             assert l.name == 'kitname'
